[DEV-5731] Move parse_request_repositories to common location
